# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Umbrella Industrial Services ERP API",
        "description": "Documentation API de l'ERP Umbrella",
        "contact": {
            "email": "support@umbrella.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://my-default-host.com",
            "description": "Serveur local"
        }
    ],
    "paths": {
...
        "/api/sites": {
            "get": {
                "tags": [
                    "Sites"
                ],
                "summary": "Liste des sites",
                "operationId": "ea87f5937c918f54f392a75c5942fb7a",
                "responses": {
                    "200": {
                        "description": "Liste des sites récupérée"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Sites"
                ],
                "summary": "Créer un site",
                "operationId": "ce626469d8f7287955203476a4c8c93f",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name",
                                    "location",
                                    "responsable_id"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "Site Centre"
                                    },
                                    "location": {
                                        "type": "string",
                                        "example": "Yaoundé"
                                    },
                                    "responsable_id": {
                                        "type": "integer",
                                        "example": 1
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Site créé avec succès"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/sites/{site}": {
            "get": {
                "tags": [
                    "Sites"
                ],
                "summary": "Afficher un site spécifique",
                "operationId": "c8b4289eb321c192034685322928bce9",
                "parameters": [
                    {
                        "name": "site",
                        "in": "path",
                        "description": "ID du site",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Site récupéré avec succès"
                    },
                    "404": {
                        "description": "Site non trouvé"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Sites"
                ],
                "summary": "Mettre à jour un site",
                "operationId": "9f2f6bb359479192c24a82c651e40876",
                "parameters": [
                    {
                        "name": "site",
                        "in": "path",
                        "description": "ID du site",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "Site Nord"
                                    },
                                    "location": {
                                        "type": "string",
                                        "example": "Douala"
                                    },
                                    "responsable_id": {
                                        "type": "integer",
                                        "example": 2
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Site mis à jour avec succès"
                    },
                    "404": {
                        "description": "Site non trouvé"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Sites"
                ],
                "summary": "Supprimer un site",
                "operationId": "51d2b5b44b105d0ff8c024c3342c1e9f",
                "parameters": [
                    {
                        "name": "site",
                        "in": "path",
                        "description": "ID du site",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Site supprimé avec succès"
                    },
                    "404": {
                        "description": "Site non trouvé"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/users": {
            "get": {
                "tags": [
                    "Users"
                ],
                "summary": "Lister les utilisateurs",
                "operationId": "ac4677cdc04c8254bc8e5b4bc1ea8607",
                "responses": {
                    "200": {
                        "description": "Liste paginée"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Users"
                ],
                "summary": "Créer un utilisateur",
                "operationId": "275321cd2f90b4f90ef6efb5d8eef292",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/UserRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Utilisateur créé"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/users/{id}": {
            "get": {
                "tags": [
                    "Users"
                ],
                "summary": "Voir un utilisateur",
                "operationId": "eb5bb5660985e3682daa6b84e616e5b0",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Détails"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Users"
                ],
                "summary": "Mettre à jour un utilisateur",
                "operationId": "212f9157f8864781d74b04eeb56b182a",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/UserRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Utilisateur mis à jour"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Users"
                ],
                "summary": "Supprimer un utilisateur",
                "operationId": "9c8fd2108d91156811259b92773e608a",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Utilisateur supprimé"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        }
    },
    "components": {
        "schemas": {
            "UserRequest": {
                "required": [
                    "first_name",
                    "last_name",
                    "location",
                    "email",
                    "password"
                ],
                "properties": {
                    "first_name": {
                        "type": "string",
                        "example": "John"
                    },
                    "last_name": {
                        "type": "string",
                        "example": "Doe"
                    },
                    "location": {
                        "type": "string",
                        "example": "Yaoundé"
                    },
                    "phone": {
                        "type": "string",
                        "example": "699123456"
                    },
                    "email": {
                        "type": "string",
                        "example": "john.doe@example.com"
                    },
                    "password": {
                        "type": "string",
                        "example": "secret123"
                    },
                    "password_confirmation": {
                        "type": "string",
                        "example": "secret123"
                    }
                },
                "type": "object"
            },
...
        }
    },
...
}